crud_api/requirements.txt
fastapi==0.112.2
omegaconf==2.3.0
SQLAlchemy==2.0.25
sqlmodel==0.0.14
pydantic==2.8.2
python-jose==3.3.0
passlib==1.7.4
uvicorn==0.20.0
bcrypt==4.1.2
asyncpg==0.29.0
python-multipart==0.0.7
psycopg2-binary==2.9.9
SQLAlchemy-Utils==0.41.2
python-dotenv==1.0.1